requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.95.0
fastapi-cache2==0.2.1
fastapi-users==10.4.1
fastapi-users-db-sqlalchemy==5.0.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
orjson==3.8.9
packaging==23.0
passlib==1.7.4
pendulum==2.1.2
pluggy==1.0.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pytest==7.3.0
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytzdata==2020.1
PyYAML==6.0
redis==4.5.4
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.9
starlette==0.26.1
timestamp==0.0.1
tomli==2.0.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
yarl==1.8.2
